requirements.txt
robobrowser>=0.5.3
bs4>=0.0.1
lxml>=4.1.0
pyodbc>=4.0.17
azure>=0.11.0
azure-storage>=0.36.0
azure-common>=1.1.5
azure-storage-common~=1.3
python-telegram-bot>=10.1.0
